style: :lipstick: Reformat code
Se organizó el proyecto en general
