+++ src/context/productoAcademicoContext.js
@@ -1,5 +1,4 @@
 import { createContext } from 'react';
-
-import { competencias } from '../../data/competencias';
+import { competencias } from '../data/competencias';
 
 export const ProductoAcademicoContext = createContext(competencias);
